# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Biobanking API",
    "version": "0.1"
  },
  "paths": {
    "/users": {
      "get": {
        "summary": "Get list of users",
        "description": "Returns a list of users",
        "responses": {
          "200": {
            "description": "A list of users",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          },
          "401": {
            "$ref": "#/responses/Standard401ErrorResponse"
          },
          "409": {
            "$ref": "#/responses/Standard409ErrorResponse"
          },
          "500": {
            "$ref": "#/responses/Standard500ErrorResponse"
          }
        }
      },
      "post": {
        "summary": "Create a user",
        "description": "Adds a new user to the user's list",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "The user to create.",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/User"
                },
                {
                  "properties": {
                    "password": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User successfully created.",
            "schema": {
              "allOf": [
                {
                  "properties": {
                    "user": {
                      "$ref": "#/definitions/User"
                    },
                    "token": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "$ref": "#/responses/Standard400ErrorResponse"
          },
          "500": {
            "$ref": "#/responses/Standard500ErrorResponse"
          }
        }
      }
    },
...
    "/users/token": {
      "post": {
        "summary": "Request token",
        "description": "Get a token using valid user credentials",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "The user to create.",
            "schema": {
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully created.",
            "schema": {
              "allOf": [
                {
                  "properties": {
                    "user": {
                      "$ref": "#/definitions/User"
                    },
                    "token": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "$ref": "#/responses/Standard400ErrorResponse"
          },
          "500": {
            "$ref": "#/responses/Standard500ErrorResponse"
          }
        }
      },
      "delete": {
        "summary": "Remove token",
        "description": "Remove a token",
        "parameters": [
          {
            "$ref": "#/parameters/jwtAuthorization"
          }
        ],
        "responses": {
          "204": {
            "description": "Token successfuly deleted."
          },
          "401": {
            "$ref": "#/responses/Standard401ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/Standard404ErrorResponse"
          },
          "409": {
            "$ref": "#/responses/Standard409ErrorResponse"
          },
          "500": {
            "$ref": "#/responses/Standard500ErrorResponse"
          }
        }
      }
    },
...
  },
  "parameters": {
    "jwtAuthorization": {
      "name": "Authorization",
      "description": "JWT token",
      "type": "string",
      "in": "header",
      "required": true
    }
  },
  "definitions": {
    "User": {
      "required": [
        "username",
        "fullname"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "readOnly": true
        },
        "username": {
          "type": "string"
        },
        "fullname": {
          "type": "string"
        },
        "status": {
          "type": "string",
          "readOnly": true
        }
      }
    },
    "Users": {
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User"
          }
        }
      }
    },
...
    "Error": {
      "properties": {
        "code": {
          "type": "string"
        },
        "errors": {
          "type": "array",
          "items": {
            "properties": {
              "field": {
                "type": "string"
              },
              "error": {
                "type": "string"
              }
            }
          }
        },
        "userMessage": {
          "type": "string"
        },
        "internalMessage": {
          "type": "string"
        }
      }
    }
  },
  "responses": {
    "Standard500ErrorResponse": {
      "description": "An unexpected error occured."
    },
    "Standard400ErrorResponse": {
      "description": "Incorrect parameters submitted.",
      "schema": {
        "$ref": "#/definitions/Error"
      }
    },
    "Standard401ErrorResponse": {
      "description": "Unauthorized"
    },
    "Standard404ErrorResponse": {
      "description": "Resource does not exists."
    },
    "Standard409ErrorResponse": {
      "description": "Your app is making too many requests and is being rate limited. 429s can trigger on a per-app or per-user basis."
    }
  }
}